feat: add conditional formats check
